start work on using a PEG-generated parser
