+++ compiler/.gitignore
@@ -1,0 +1,2 @@
+parser.c
+./include/parser.h
